partial combination of master (UI) with weapons branch

+++ scripts/damage_system/Hitbox.cs
@@ -6,6 +6,6 @@ using Godot;
 
 public partial class Hitbox : Area2D
 {
-	[Export] public int Damage = 10;
-	[Export] public int KnockbackWeight = 800;
+	[Export] public float Damage = 10;
+	[Export] public float KnockbackWeight = 800;
 }
