+++ scripts/enemy_state_machines/base_classes/EInstanceContext.cs
@@ -5,5 +5,5 @@ namespace GameJam25.scripts.enemy_state_machines.base_classes;
 public partial class EInstanceContext : Node2D
 {
     public Vector2 KnockbackDir = Vector2.Zero;
-    public int KnockbackWeight = 0;
+    public float KnockbackWeight = 0;
 }
